configure: Adds wayland config options
This adds a --enable-wayland configure option to enable building mutter
as a hybrid X and Wayland compositor. By default the option is disabled.
If enabled HAVE_WAYLAND is defined for C code and as an automake
conditional.

In addition a --with-wayland-protocols option has been added since
wayland support will depend on the xserver.xml extension protocol that's
currently part of the Weston project which we need the location of.