Adds a --nested option
This adds a --nested option to request that mutter no longer run as a
classic X compositor with an output window mapped on the X Composite
Overlay Window and also not assume it is running directly under X.

The intention is that in this mode Mutter will itself launch a headless
X server and display output will be handled by Clutter and Cogl. This
will enable running Mutter nested as an application within an X session.

This patch introduces the internal idea that mutter is running as a
"display server" as a means to condition the way mutter operates in
various ways throughout the code base.

Later we also expect to add a --kms option as another way of enabling
this display server mode that will assume full control of the display
hardware instead of running as a nested application.